+++ packages/vue-lighthouse-viewer/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.1.10 (2021-05-18)
+
+**Note:** Version bump only for package vue-lighthouse-viewer
+
+
+
+
+
+## 0.1.9 (2021-05-18)
+
+**Note:** Version bump only for package vue-lighthouse-viewer
+
+
+
+
+
+## 0.1.8 (2021-05-18)
+
+**Note:** Version bump only for package vue-lighthouse-viewer
+
+
+
+
+
 ## [0.1.7](https://github.com/dvelasquez/lighthouse-viewer/compare/vue-lighthouse-viewer@0.1.6...vue-lighthouse-viewer@0.1.7) (2021-05-09)
 
 **Note:** Version bump only for package vue-lighthouse-viewer